requirements.txt
Flask
SpeechRecognition
anthropic
python-dotenv
Werkzeug